Revert "Update swagger-ui to 2.2.10"
It turns out that the swagger guys have fixed the bug that meant that JSON
examples were interpreted as json encoded strings rather than JSON, which means
that our JSON is now all broken.

For now, let's roll back to our old version of the swagger UI. When/if we
release the spec, we can fix the json examples at the same time.