Fix the RSS feed so it shows the full post (#408)
Apparently the html output was just broken on that version of gatsby-mdx. When I update it to version 0.6.2 the RSS feed works but if I update it to the latest version (0.6.3) I get some awful react error when building the site. No idea up with that but at least the RSS feed works.

Also had to remove a random character from one of the blog posts to get the XML to be valid. We may need some way of validating blog posts so those characters don't get added and silently break the feed.

Signed-off-by: Aaron Raimist <aaron@raim.ist>