Added my French user introduction guide
I've been writing a guide for newcomers in French. It describes what is Matrix (with a clear focus on escaping surveillance capitalism) and how to install/register to Riot. It's the front page of https://zinz.dev: my own Synapse instance.
I plan on adding an "how to make friends" section, and to translate the guide in English. I will keep you posted.
Hope it helps!