Freeze the current appendices.html file for the legacy spec (#964)
This PR is part of unfreezing the spec as a whole after the spec redesign has launched: https://github.com/matrix-org/matrix-doc/issues/3055

Essentially we'd like to redirect the unstable versions of the old spec to the new spec site (https://spec.matrix.org). Unfortunately, we can't do that for the appendices page as there is no stable vs. unstable appendices page. It's never been versioned. As such, if we redirected it then any links to the appendices in the stable versions of the legacy spec would take you to the appendices page on the new spec - which would be very weird.

Thus, this PR freezes the appendices page in time. We'll be setting up redirects for all of the `/unstable` spec pages, as well as the `/proposals` page, as with the former the user consciously chooses to view the unstable spec, and in the latter the proposals page is a living document which isn't deep-linked into by the spec.