Add anchor links
See https://www.getzola.org/documentation/content/section/#front-matter
after the configs, it writes:

"Keep in mind that any configuration options apply only to the direct
pages, not to the subsections' pages."

So it must be applied to all _index.md for it to be everywhere.

I have not applied the same thing for folders without a `_index.md`
file, and anchor links doesn't seem to be applied to `<h1>` elements
(maybe within `<header>`). It's possible I may have missed something.

fixes #1760