Fix vertical centering of the matrix logo on matrix.org
In classical graphical design you almost never center something to literally 50% of the height of the screen; instead you aim for the golden ratio of 0.618:1 rather than 1:1.  I've tweaked this a bit more to account for the visual density of the logo & accompanying text.  Closes https://github.com/matrix-org/matrix.org/issues/1836