Remove `overflow: hidden` on page headers
It shouldn't be possible for the contents of page headers to overflow, considering the page header is bound by a variable sized bounding box (i.e., `min-height` instead of `height`).

I think it's dangerous to have `overflow: hidden` without reason, as if the design's implementation were to change, we would could be cutting off content without realising it.