Allow setting author metadata and merge author aliases
This allows to set a canonical name of people which have multiple aliases used as authors. It will keep links intact and just merge the posts as needed.
It also updates the links in the posts to the correct name and slug.
Additionally this feature allows for a description on the author page to exist to introduce a person.
Last but not least this allows setting a mastaodon uri to allow using the fediverse:creator meta tag