Add official account info page and client screenshots
Introduces a new documentation page explaining the Matrix.org (Official Account) user, how to identify it, and what to do if you receive suspicious messages. Includes example screenshots for different clients to help users verify the official account.

The purpose is to support using Server Notices more extensively on the matrix.org homeserver. This new page will be referred in the topic of the server notices room.

n.b currently the server notices room is titled Server Alerts and has a different avatar, but these will be changed at the same time as putting this page live.