Introduce different usage limits for Free and Premium account plans
This PR is part of the project to introduce premium accounts to fund the matrix.org homeserver. See https://matrix.org/blog/2025/06/funding-homeserver-premium/

We are introducing a lower data usage limit for accounts on the Free plan (100MB/24 hours + 500MB/28 days) and a higher data usage limit for accounts on the Premium plan (1GB/24 hours + 5GB/28 days).

The merging of this PR needs to be coordinated with the rollout of the necessary configuration changes.